+++ src/pages/ClassesPage.jsx
@@ -4,7 +4,8 @@ import { useLocation, useParams, Link } from 'wouter';
 import Class from '../components/Class/Class';
 import Level from '@/components/Class/Level';
 import SkeletonClass from '@/components/Skeletons/SkeletonClass';
-import { getClasses, getLevels } from '../api/class-wrapper';
+import { getClasses } from '@/api/class-wrapper';
+import { getLevels } from '@/api/level-wrapper';
 import { useTranslation } from "react-i18next";
 import SkeletonLevel from '../components/Skeletons/SkeletonLevel';
 import Skeleton from 'react-loading-skeleton';
