+++ src/pages/dashboards/admin/editPages/EditLevel.jsx
@@ -3,7 +3,8 @@ import { UserContext } from '@/contexts/UserContext.jsx';
 import { useLocation, useParams } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
 import { useTranslation } from "react-i18next";
-import { getLevels, updateLevel, deleteLevel, getClasses, deleteClass } from '@/api/class-wrapper.js';
+import { getLevels, updateLevel, deleteLevel } from '@/api/level-wrapper.js';
+import { getClasses, deleteClass } from '@/api/class-wrapper.js';
 import Button from '@/components/Button/Button';
 import DeleteButton from "@/components/Button/DeleteButton";
 import BackButton from "@/components/Button/BackButton";
